# Improve bundling speed replacing inefficient reduce code