@ryan-roemer 2020-06-23
Looks great and thanks for the write up!

I don't know why CI didn't kick in. I'm going to close and then re-open this PR and see if that starts it....